requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycryptodome==3.9.8
pyparsing==2.4.7
pytest==5.4.3
pytest-flask==1.0.0
pytest-mock==3.2.0
regex==2020.7.14
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.1.0
